+++ kernel/kernel-api/src/test/scala/io/delta/kernel/utils/ExpressionTestUtils.scala
@@ -1,21 +1,6 @@
-/*
- * Copyright (2023) The Delta Lake Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.delta.kernel.defaults.utils
+package io.delta.kernel.utils
 
-import io.delta.kernel.expressions.{Column, Expression, Literal, Predicate}
+import io.delta.kernel.expressions.{Column, Expression, Predicate}
 
 /** Useful helper functions for creating expressions in tests */
 trait ExpressionTestUtils {
